Add CC icons, fix / improve spelling & wording
First and foremost, I had the desire to add the Creative Commons license of the particular license I am using to the copyright.tex file — and have done so using a very helpful package. Additionally, there were a handful of grammatical and spelling errors that I corrected in chapter fifteen and the introductory text.