Add "note" chapter, change copyright, change chapter headings, better URLs
A rather large number of changes in this commit. I have added a "note" chapter, wherein I explain how one may obtain a copy of my book and compile and how it can be used in accordance to the license terms. Additionally, I adjusted the size of the image in the last chapter so that it now fills the entire page. I have also amended the disclaimer in the "copyright" section, so that it now also tells the reader that the Ancient Greek text is in the public domain as well as the illustrations. Another thing I was greatly annoyed by were the very strangely formatted URLs, so I changed the way they are displayed as well.